% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anon.R
\name{anonymise}
\alias{anonymise}
\title{Anonymise factor columns across datasets in a consistent way}
\usage{
anonymise(data_list, prefix = "", return_original_levels = FALSE)
}
\arguments{
\item{data_list}{A list of data frames or tibbles.}

\item{prefix}{A character prefix to insert in front of the random labels.}

\item{return_original_levels}{Whether or not the resulting list should also include the original, non-anonymised levels. Default: FALSE.}
}
\value{
A list containing the original data, but with consistently anonymised factors
}
\description{
`anonymise()` is a useful function for anonymising factor columns
across different datasets using consistent anonymised levels. In other
words, if the same factor level appears in more than one dataset, then
`anonymise()` will use the same anonymous factor for that level.
}
